Matching behavior with pandas 1.1.0

+++ databricks/koalas/tests/test_series.py
@@ -1284,18 +1284,20 @@ class SeriesTest(ReusedSQLTestCase, SQLTestUtils):
         self.assert_eq(pser1.mask(pser1 > 3), kser1.mask(kser1 > 3).sort_index(), almost=True)
 
     def test_truncate(self):
-        pser1 = pd.Series([10, 20, 30, 40, 50, 60, 70], index=[1, 2, 3, 4, 5, 6, 7])
-        kser1 = ks.Series(pser1)
-        pser2 = pd.Series([10, 20, 30, 40, 50, 60, 70], index=[7, 6, 5, 4, 3, 2, 1])
-        kser2 = ks.Series(pser2)
+        # The bug has been fixed in pandas 1.1.0.
+        if LooseVersion(pd.__version__) >= LooseVersion("1.1.0"):
+            pser1 = pd.Series([10, 20, 30, 40, 50, 60, 70], index=[1, 2, 3, 4, 5, 6, 7])
+            kser1 = ks.Series(pser1)
+            pser2 = pd.Series([10, 20, 30, 40, 50, 60, 70], index=[7, 6, 5, 4, 3, 2, 1])
+            kser2 = ks.Series(pser2)
 
-        self.assert_eq(kser1.truncate(), pser1.truncate())
-        self.assert_eq(kser1.truncate(before=2), pser1.truncate(before=2))
-        self.assert_eq(kser1.truncate(after=5), pser1.truncate(after=5))
-        self.assert_eq(kser1.truncate(copy=False), pser1.truncate(copy=False))
-        self.assert_eq(kser1.truncate(2, 5, copy=False), pser1.truncate(2, 5, copy=False))
-        self.assert_eq(kser2.truncate(4, 6), pser2.truncate(4, 6))
-        self.assert_eq(kser2.truncate(4, 6, copy=False), pser2.truncate(4, 6, copy=False))
+            self.assert_eq(kser1.truncate(), pser1.truncate())
+            self.assert_eq(kser1.truncate(before=2), pser1.truncate(before=2))
+            self.assert_eq(kser1.truncate(after=5), pser1.truncate(after=5))
+            self.assert_eq(kser1.truncate(copy=False), pser1.truncate(copy=False))
+            self.assert_eq(kser1.truncate(2, 5, copy=False), pser1.truncate(2, 5, copy=False))
+            self.assert_eq(kser2.truncate(4, 6), pser2.truncate(4, 6))
+            self.assert_eq(kser2.truncate(4, 6, copy=False), pser2.truncate(4, 6, copy=False))
 
         kser = ks.Series([10, 20, 30, 40, 50, 60, 70], index=[1, 2, 3, 4, 3, 2, 1])
         msg = "truncate requires a sorted index"
